backend/requirements.txt
fastapi==0.118.0
gunicorn==23.0.0
uvicorn[standard]==0.37.0
sqlalchemy==2.0.43
alembic==1.14.0
pydantic-settings==2.11.0
psycopg[binary]==3.2.10
python-jose==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pytest==8.4.2
httpx==0.28.1
python-multipart==0.0.20
redis==5.2.1
prometheus-client==0.21.1
